refactor: modularize FitsAnalyzer into a package
- Decomposed apts/utils/fits_analyzer.py into apts/utils/fits_analyzer/ package.
- Extracted I/O logic to loader.py.
- Extracted star fitting and detection to fitter.py.
- Extracted backfocus classification and surface building to classification.py.
- Centralized constants to constants.py to prevent circular imports.
- Updated FitsAnalyzer in base.py to delegate to the new modules while preserving the public API.
- Updated unit tests to reflect the new internal module structure for mocks.
