+++ apts/utils/fits_analyzer/__init__.py
@@ -1,0 +1,4 @@
+from .base import FitsAnalyzer
+from .constants import ANALYSIS_DEFAULTS
+
+__all__ = ["FitsAnalyzer", "ANALYSIS_DEFAULTS"]
