+++ apts/utils/fits_analyzer/constants.py
@@ -1,0 +1,23 @@
+ANALYSIS_DEFAULTS = {
+    "fwhm_est": 5.0,
+    "threshold": 8.0,
+    "star_limit": 500,
+    "edge_margin": 50,
+    "box_size": 25,
+    "min_stars": 10,
+    "retry_threshold": 4.0,
+    "max_eccentricity": 0.95,
+    "max_chi2": 5.0,
+    "progress_interval": 20,
+    "ecc_threshold": 0.15,
+    "radial_positive_threshold": 0.3,
+    "radial_negative_threshold": -0.3,
+    "annular_zones": 4,
+    "annular_weights": [0.1, 0.2, 0.3, 0.4],
+    "grid_size": 50,
+    "poly_degree": 2,
+    "sigma_clip_iters": 3,
+    "sigma_clip_sigma": 2.5,
+    "autobin_threshold": 4096,
+    "mosaic_tile_fraction": 0.20,
+}
